Move on-chain network parameters to actors/builtin.
Also remove use of ipld stub library.
