+++ src/actors/runtime/runtime.id
@@ -3,7 +3,6 @@ import abi "github.com/filecoin-project/specs/actors/abi"
 import exitcode "github.com/filecoin-project/specs/actors/runtime/exitcode"
 import addr "github.com/filecoin-project/go-address"
 import indices "github.com/filecoin-project/specs/actors/runtime/indices"
-import ipld "github.com/filecoin-project/specs/libraries/ipld"
 import cid "github.com/ipfs/go-cid"
 
 // Runtime is the VM's internal runtime object.
@@ -113,9 +112,9 @@ type Runtime interface {
     DeleteActor(address addr.Address)
 
     // Retrieves and deserializes an object from the store into o. Returns whether successful.
-    IpldGet(c cid.Cid, o ipld.Object) bool
+    IpldGet(c cid.Cid, o interface {}) bool
     // Serializes and stores an object, returning its CID.
-    IpldPut(x ipld.Object) cid.Cid
+    IpldPut(x interface {}) cid.Cid
 
     // Provides the system call interface.
     Syscalls() Syscalls
