+++ src/systems/filecoin_blockchain/struct/chain/chain.go
@@ -2,6 +2,7 @@ package chain
 
 import (
 	abi "github.com/filecoin-project/specs/actors/abi"
+	builtin "github.com/filecoin-project/specs/actors/builtin"
 	node_base "github.com/filecoin-project/specs/systems/filecoin_nodes/node_base"
 )
 
@@ -26,9 +27,9 @@ func (chain *Chain_I) GetTicketProductionRandSeed(epoch abi.ChainEpoch) abi.Rand
 }
 
 func (chain *Chain_I) GetSealRandSeed(epoch abi.ChainEpoch) abi.RandomnessSeed {
-	return chain.RandomnessAtEpoch(epoch - node_base.SPC_LOOKBACK_SEAL)
+	return chain.RandomnessAtEpoch(epoch - builtin.SPC_LOOKBACK_SEAL)
 }
 
 func (chain *Chain_I) GetPoStChallengeRandSeed(epoch abi.ChainEpoch) abi.RandomnessSeed {
-	return chain.RandomnessAtEpoch(epoch - node_base.SPC_LOOKBACK_POST)
+	return chain.RandomnessAtEpoch(epoch - builtin.SPC_LOOKBACK_POST)
 }
